+++ source/Blueprint/Reflection/EnumType.hpp
@@ -9,10 +9,10 @@ namespace reflection
     class EnumType : public Type
     {
     public:
-        using Entry = std::pair<std::string, int>;
+        using Entry = std::pair<std::string, size_t>;
         using Entries = std::vector<Entry>;
 
-        void AddEntry(const std::string& name, int value);
+        void AddEntry(const std::string& name, size_t value);
         const Entries& GetEntries() const;
 
         virtual void Accept(TypeVisitor& visitor) const override;
